fix(os): make lv_lock/unlock public as they should be used by the users
