+++ lvgl.h
@@ -37,6 +37,8 @@ extern "C" {
 #include "src/misc/lv_circle_buf.h"
 #include "src/misc/lv_tree.h"
 
+#include "src/osal/lv_os_private.h"
+
 #include "src/tick/lv_tick.h"
 
 #include "src/core/lv_obj.h"
