+++ src/draw/sw/lv_draw_sw.h
@@ -19,7 +19,7 @@ extern "C" {
 #include "../../misc/lv_area.h"
 #include "../../misc/lv_color.h"
 #include "../../display/lv_display.h"
-#include "../../osal/lv_os.h"
+#include "../../osal/lv_os_private.h"
 
 #include "../lv_draw_vector.h"
 #include "../lv_draw_triangle.h"
