+++ src/draw/sw/lv_draw_sw_mask.c
@@ -15,7 +15,7 @@
 #include "../../misc/lv_math.h"
 #include "../../misc/lv_log.h"
 #include "../../misc/lv_assert.h"
-#include "../../osal/lv_os.h"
+#include "../../osal/lv_os_private.h"
 #include "../../stdlib/lv_string.h"
 
 /*********************
