+++ src/libs/vg_lite_driver/lv_vg_lite_hal/lv_vg_lite_hal.c
@@ -2,7 +2,7 @@
 #include "../../../lv_conf_internal.h"
 #if LV_USE_VG_LITE_DRIVER
 
-#include "../../../osal/lv_os.h"
+#include "../../../osal/lv_os_private.h"
 #include "../../../stdlib/lv_mem.h"
 
 #if LV_USE_OS == LV_OS_NONE
