+++ src/osal/lv_cmsis_rtos2.c
@@ -12,7 +12,7 @@
 /*********************
  *      INCLUDES
  *********************/
-#include "lv_os.h"
+#include "lv_os_private.h"
 
 #if LV_USE_OS == LV_OS_CMSIS_RTOS2
