+++ src/osal/lv_linux.c
@@ -7,7 +7,7 @@
  *      INCLUDES
  *********************/
 
-#include "lv_os.h"
+#include "lv_os_private.h"
 
 #if LV_USE_OS != LV_OS_NONE && defined(__linux__)
