+++ src/osal/lv_os_none.c
@@ -10,7 +10,7 @@
 #include "../lv_conf_internal.h"
 #if LV_USE_OS == LV_OS_NONE
 
-#include "lv_os.h"
+#include "lv_os_private.h"
 #include "../misc/lv_timer.h"
 
 /*********************
